requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
pymongo==4.6.1
Werkzeug==3.0.1
python-dotenv==1.0.0
dnspython==2.4.2
razorpay==1.4.1
setuptools==70.0.0
Flask-Mail==0.9.1
requests==2.31.0
oauthlib==3.2.2
pyOpenSSL==24.0.0
